Completed 53

+++ lab/esercizi/sett_5/53-implement-with-collections/src/main/java/it/unibo/collections/design/impl/ProductImpl.java
@@ -6,7 +6,7 @@ public class ProductImpl implements Product {
     private final String name;
     private final double quantity;
 
-    public ProductImpl(final String name, final int quantity) {
+    public ProductImpl(final String name, final double quantity) {
         this.name = name;
         this.quantity = quantity;
     }
